+++ refgenconf/__init__.py
@@ -3,3 +3,23 @@ from .refgenconf import RefGenConf, upgrade_config
 from .helpers import select_genome_config, get_dir_digest
 from .const import *
 from .exceptions import *
+from .helpers import *
+from .refgenconf import *
+from .populator import looper_refgenie_plugin
+
+__all__ = (
+    [
+        "RefGenConf",
+        "select_genome_config",
+        "get_dir_digest",
+        "GenomeConfigFormatError",
+        "MissingAssetError",
+        "MissingConfigDataError",
+        "MissingGenomeError",
+        "RefgenconfError",
+        "UnboundEnvironmentVariablesError",
+    ]
+    + ["DEFAULT_SERVER"]
+    + CFG_KEY_NAMES
+    + ["looper_refgenie_plugin"]
+)